refactor: change folder names
