+++ src/app/app.module.ts
@@ -6,10 +6,10 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CinemahallModule } from './pages/cinemahall/cinemahall.module';
-import { ApiKeyDatabaseInterceptor } from './servises/http/api-key-database.interceptor';
+import { ApiKeyDatabaseInterceptor } from './infrastructure/http/api-key-database.interceptor';
 import { MediaDetailsModule } from './pages/media-details/media-details.module';
 import { PersonalDetailsModule } from './pages/personal-details/personal-details.module';
-import { ApiUrlDatabaseInterceptor } from './servises/http/api-url-database.interceptor';
+import { ApiUrlDatabaseInterceptor } from './infrastructure/http/api-url-database.interceptor';
 
 @NgModule({
   declarations: [
