+++ src/app/infrastructure/utiles/request.ts
@@ -1,5 +1,5 @@
 import { forkJoin, Observable} from "rxjs"
-import { MediaType } from "src/app/models/media/types"
+import { MediaType } from "src/app/features/media/types"
 
 
 export function mergeRequest<RequestResult>(dataList: MediaType[], requestFn: Function): Observable<RequestResult[]> {
