+++ src/app/pages/media-details/media-details.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { MediaDetailsComponent } from './media-details.component';
 
 
@@ -11,8 +10,6 @@ import { MediaDetailsComponent } from './media-details.component';
   exports: [
     MediaDetailsComponent
   ],
-  imports: [
-    CommonModule
-  ]
+  imports: []
 })
 export class MediaDetailsModule { }
